Cloud_Functions/BigQuery_loader_fazenda_CF/requirements.txt
google-cloud-bigquery==3.*
functions-framework==3.*

